feat: initial windows port
