+++ go.mod
@@ -3,7 +3,8 @@ module filippo.io/yubikey-agent
 go 1.14
 
 require (
+	github.com/Microsoft/go-winio v0.4.14
 	github.com/go-piv/piv-go v1.5.0
-	github.com/gopasspw/gopass v1.9.1
+	github.com/gopasspw/gopass v1.9.3-0.20200526132452-b2f36ba7fb8a
 	golang.org/x/crypto v0.0.0-20200429183012-4b2356b1ed79
 )
