+++ impl_unix.go
@@ -1,0 +1,39 @@
+// +build darwin dragonfly freebsd linux netbsd openbsd
+
+package main
+
+import (
+	"log"
+	"net"
+	"os"
+	"os/signal"
+	"syscall"
+
+	"golang.org/x/crypto/ssh/terminal"
+)
+
+func runAgent(socketPath string) {
+	if terminal.IsTerminal(int(os.Stdin.Fd())) {
+		log.Println("Warning: yubikey-agent is meant to run as a background daemon.")
+		log.Println("Running multiple instances is likely to lead to conflicts.")
+		log.Println("Consider using the launchd or systemd services.")
+	}
+
+	a := &Agent{}
+
+	c := make(chan os.Signal)
+	signal.Notify(c, syscall.SIGHUP)
+	go func() {
+		for range c {
+			a.Close()
+		}
+	}()
+
+	os.Remove(socketPath)
+	l, err := net.Listen("unix", socketPath)
+	if err != nil {
+		log.Fatalln("Failed to listen on UNIX socket:", err)
+	}
+
+	serveConns(l, a)
+}
